Revamped multi_array

+++ src/core/vec3.h
@@ -194,7 +194,7 @@ struct Extent : public Vec3<int> {
   HD_INLINE int& depth() { return z; }
   HD_INLINE const int& depth() const { return z; }
 
-  HD_INLINE int size() const { return x * y * z; }
+  HD_INLINE size_t size() const { return x * y * z; }
 
   // template <typename T>
   // cudaExtent cuda_ext(const T& t) const {
